requirements.txt

streamlit>=1.37.0
pillow>=10.4.0
bcrypt>=4.1.3
pandas>=2.2.2
matplotlib>=3.8.0
fpdf>=1.7.2
python-barcode>=0.14
reportlab>=4.0.0
gspread>=5.8.0
oauth2client>=4.1.3
qrcode>=7.4
pytest>=7.4.0
